@@ -1,11 +1,11 @@
 name: objectid
 description: Blazing fast, cross-platform ObjectId implementation for the dart language.
-version: 1.1.1
+version: 2.0.0-nullsafety.0
 homepage: https://github.com/gonuit/dart_objectid
 repository: https://github.com/gonuit/dart_objectid
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0-0 <3.0.0"
 
 dev_dependencies:
   effective_dart: ^1.2.0
